Move relevant utils.h
